requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.4
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-tables2==2.7.5
idna==3.10
jsonfield==3.1.0
pillow==11.2.1
python-decouple==3.8
python-dotenv==0.21.1
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
urllib3==2.4.0
